backend/requirements.txt
flask
flask-cors
numpy
scikit-learn
joblib
gunicorn
google-generativeai
python-dotenv